requirements.txt
torch==2.0.1
torchvision>=0.15.0
transformers==4.30.2
opencv-python>=4.7.0
pytesseract>=0.3.10
fastapi>=1.34.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
kagglehub>=0.1.0
